Split up `font` module
